Rework gis view

+++ common.moplaf.gisview/com.misc.common.moplaf.gisview/src/com/misc/common/moplaf/gisview/ILocationProvider.java
@@ -12,8 +12,8 @@ package com.misc.common.moplaf.gisview;
 
 
 public interface ILocationProvider {
-	float getLongitude(Object element);
-	float getLatitude(Object element);
-	float getElevation(Object element);
+	double getLongitude(Object element);
+	double getLatitude(Object element);
+	double getElevation(Object element);
 	boolean isLocation(Object element);
 }
